+++ src/org/usfirst/frc/team5188/robot/Shooter.java
@@ -9,7 +9,7 @@ public class Shooter implements PID_Actuator, PID_Sensor {
 	public Pats_PID_Controller controller;
 	public double lastSet;
 	
-	Shooter(int motor_pin, double p, double i, double d, int loopTime, boolean inverted){
+	Shooter(int motor_pin, double p, double i, double d, double loopTime, boolean inverted){
 		talon = new CANTalon(1);
 		controller = new Pats_PID_Controller(p, i, d, loopTime, this, this);
 	}
